+++ packages/ui-theme/icloud-theme/scripts/build.sh
@@ -1,17 +1,24 @@
-echo "node $(node -v)"
-echo "npm $(npm -v)"
-echo "pnpm $(pnpm -v)"
-
-rm -rf dist
+rm -rf es
+rm -rf dist 
 rm -rf vars
-mkdir -p dist/theme
-mkdir -p vars
 
-tsc
-echo 'tsc done'
+
+tsc -p .
+
+swc variables -d es/theme --source-maps
+# 把.d.ts文件复制到es/theme目录下
+tsc --emitDeclarationOnly --declaration --outDir es/theme
+
+mkdir -p dist/theme
+
 node scripts/build.mjs
 
+cp -r patches/* es
 cp -r patches/* dist
+
+# 需要把替换好的css变量复制到es里
+cp -r vars/*.css es/theme
 cp -r vars/* dist/theme
+
 
 echo "build success"