+++ packages/ui-theme/icloud-theme/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0-beta.10](https://gitee.com/gitee-fe/osui/tree/master/compare/v4.0.0-beta.8...v4.0.0-beta.10) (2025-10-30)
+
+
+### Bug Fixes
+
+* icloud-theme build ([a81ca1a](https://gitee.com/gitee-fe/osui/tree/master/commits/a81ca1aa604129708301279ed8eb2a3976f078fe))
+* 修复icloud-theme打包问题,修复space导出问题 ([5ef687b](https://gitee.com/gitee-fe/osui/tree/master/commits/5ef687b999b1f3eb4e6a589a2b407e9650947247))
+* 增加颜色变量line-color ([0c96f27](https://gitee.com/gitee-fe/osui/tree/master/commits/0c96f27f4207f75925575e7a380a88cb151ecfb2))
+
+
+
+
+
+# [4.0.0-beta.9](https://gitee.com/gitee-fe/osui/tree/master/compare/v4.0.0-beta.8...v4.0.0-beta.9) (2025-10-30)
+
+
+### Bug Fixes
+
+* 修复icloud-theme打包问题,修复space导出问题 ([5ef687b](https://gitee.com/gitee-fe/osui/tree/master/commits/5ef687b999b1f3eb4e6a589a2b407e9650947247))
+* 增加颜色变量line-color ([0c96f27](https://gitee.com/gitee-fe/osui/tree/master/commits/0c96f27f4207f75925575e7a380a88cb151ecfb2))
+
+
+
+
+
 # [4.0.0-beta.8](https://gitee.com/gitee-fe/osui/tree/master/compare/v3.1.0...v4.0.0-beta.8) (2025-10-28)